# lib/db/src/schema/scripts.ts
import { pgTable, serial, text, boolean, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const scriptsTable = pgTable("scripts", {
  id: serial("id").primaryKey(),
  topic: text("topic").notNull(),
  platform: text("platform").notNull(),
  tone: text("tone").notNull(),
  duration: text("duration").notNull(),
  script: text("script").notNull(),
  hook: text("hook").notNull(),
  hashtags: text("hashtags").array().notNull().default([]),
  saved: boolean("saved").notNull().default(false),
  createdAt: timestamp("created_at").notNull().defaultNow(),
});

export const insertScriptSchema = createInsertSchema(scriptsTable).omit({ id: true, createdAt: true });
export type InsertScript = z.infer<typeof insertScriptSchema>;
export type Script = typeof scriptsTable.$inferSelect;